[SPARK-17270][SQL] Move object optimization rules into its own file (branch-2.0)
